+++ electron/preload.js
@@ -1,0 +1,2 @@
+const { ipcRenderer } = require('electron');
+window.ipcRenderer = ipcRenderer;
