+++ package.json
@@ -9,7 +9,7 @@
     "bson": "^4.2.2",
     "concurrently": "^5.2.0",
     "d3": "^5.16.0",
-    "electron": "^9.1.1",
+    "electron": "^9.4.3",
     "electron-builder": "^22.8.0",
     "electron-is-dev": "^1.2.0",
     "express": "^4.17.1",
@@ -23,9 +23,11 @@
   },
   "proxy": "http://localhost:8080",
   "scripts": {
-    "ebuild": "npm run build && node_modules/.bin/build",
-    "start": "react-scripts start",
-    "build": "react-scripts --max_old_space_size=4096 build",
+    "start": "export BROWSER=none && react-scripts start",
+    "start-electron": "export ELECTRON_START_URL=http://localhost:3000 && electron .",
+    "build": "react-scripts --max_old_space_size=8000 build",
+    "build-electron": "mkdir build/src && cp -r electron/. build/electron && cp -r src/shared/. build/src/shared",
+    "package": "electron-builder build --mac --win -c.extraMetadata.main=build/electron/main.js --publish never",
     "dev": "concurrently \"npm start\" \"wait-on http://localhost:3000 && electron .\"",
     "server": "nodemon server.js"
   },
@@ -43,5 +45,20 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "build": {
+    "files": [
+      "build/**/*",
+      "node_modules/**/*"
+    ],
+    "directories":{
+      "buildResources": "assets"
+    },
+    "electronVersion": "4.2.8",
+    "publish": {
+      "provider": "github",
+      "repo": "Nimbus",
+      "owner": "clima"
+    }
   }
 }
