+++ server.js
@@ -18,6 +18,10 @@ app.get('/dbMetadataList', (req, res) => {
 
 
 // --------------------------------------------------------
+
+
+
+// --------------------------------------------------------
 app.get('/simDiagnosticFile', (req, res) => {
   let sim = JSON.parse(req.query.sim);
   let diagPath = `${ userPath }/${ sim['site_id'] }/${ sim['sim_id'] }/_diagnostic.json`;
