+++ src/shared/constants.js
@@ -1,0 +1,5 @@
+module.exports = {
+  channels: {
+    APP_INFO: 'app_info',
+  },
+};
